NOT WORKING // Save stimulus & config in distribution in before_start then update with results on finish
The idea was to already create a distribution in s3 for a simulation as
soon as we get a request to launch a simulation.
This distribution would later be updated by the celery task to also
include simulation result. However, it looks like it's not possible to
update a file stored in s3 (see https://bluebrainproject.slack.com/archives/G013PKBUHT2/p1728567806810799)
so I will create a distribution at the end, when the simulation is
complete. This distribution will contain the simulation config,
simulation result as well as stimulus plot data.