Run simulation in an isolated process to avoid leaking simulation state
Neuron simulator does not reset itself automatically after a call to
simulation.run() This causes in-correct simulation data because it (the
data) also contains information from previous simulations that ran on
the same worker. The best way to avoid this (as suggested by neuron
forums) is to run the simulation inside the child process.

Since we cannot use multiprocessing module with celery, billiard is
used.

See:
- https://www.neuron.yale.edu/phpBB/viewtopic.php?t=4039
- https://stackoverflow.com/questions/54858326/python-multiprocessing-billiard-vs-multiprocessing
- https://github.com/celery/celery/issues/5362#issuecomment-466862879

Signed-off-by: Dinika Saxena <dinikasaxenas@gmail.com>